move more serverless & api-client boto3 functions to shared library. also setup sepertate logging & rich print for each respectively. get labels from rekognition in serverless call
