+++ api-client-sig4/client_launcher.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3.12
 
-
 import argparse
 import os
 import random
