+++ api-client-sig4/helpers/cat_api_client.py
@@ -34,14 +34,12 @@ import random
 import sys
 
 import requests
+from boto3_client_helpers import (
+    upload_local_image_2rekog_blocking,
+)
 from helpers.aws_request_signer import AWSRequestSigner
 from helpers.boto3_bulk_s3_uploader import BulkS3Uploader
 from rich import print
-
-from shared_helpers.boto3_helpers import (
-    fetch_values_from_ssm,
-    upload_local_image_2rekog_blocking,
-)
 
 
 class CatAPIClient:
