+++ api-client-sig4/helpers/config.py
@@ -27,10 +27,9 @@ Dependencies:
 import os
 import sys
 
+from boto3_client_helpers import fetch_values_from_ssm
 from dotenv import load_dotenv
 from rich import print
-
-from shared_helpers.boto3_helpers import fetch_values_from_ssm
 
 secret_vars = [
     "AWS_ACCESS_KEY_ID",
