--- api-client-sig4/pyproject.toml
@@ -1,29 +1,0 @@
-[project]
-authors = [
-  {name = "Darkpandarts", email = "14529342+darkpandarts@users.noreply.github.com"},
-]
-dependencies = []
-description = "API client that makes calls to ice-cat-wrangler API"
-name = "ice-cat-wrangler-api-client-sig4"
-readme = "README.md"
-requires-python = ">=3.12"
-version = "0.3.0"
-
-[tool.poetry]
-packages = [{include = "", from = "helpers"}]
-
-[tool.isort]
-include_trailing_comma = true
-known_first_party = ["shared"]
-line_length = 88
-multi_line_output = 3
-profile = "black"
-
-[tool.black]
-line_length = 88
-
-[build-system]
-# build-backend = "poetry.core.masonry.api"
-build-backend = "setuptools.build_meta"
-requires = ["setuptools", "wheel"]
-requires = ["poetry-core==1.8.5", "setuptools", "wheel"]
