+++ api-client-sig4/requirements.txt
@@ -1,4 +1,3 @@
-shared_helpers @ file://../shared_helpers
 certifi==2025.1.31
 charset-normalizer==3.4.1
 dotmap==1.3.30
@@ -11,3 +10,4 @@ rich==14.0.0
 ruamel.yaml==0.18.10
 ruamel.yaml.clib==0.2.12
 urllib3==2.3.0
+python-dotenv==1.1.0
